requirements.txt
transformers==4.40.1
huggingface_hub==0.23.0
datasets
torchaudio
librosa
soundfile
torch
numpy
scipy
sounddevice
jiwer
ffmpeg-python
